fix(wrapper)!: remove ability to re-initialize all generator types after construction via the seeds setter
BREAKING CHANGE: `seeds` property no longer allows setting (re-setting) seeds after RandomGenerator creation. This prevents issues with inconsistent state advancement patterns, particularly for the PCG generator which requires a stream increment (if provided) to be provided in an atomic operation with seeds. AS consumers must respect this ordering, as it is not enforced there.